Split post 1422 oversized slices
