+++ app/src/test/java/com/pocketshell/app/composer/PromptComposerViewModelTest.kt
@@ -1,7 +1,6 @@
 package com.pocketshell.app.composer
 
 import androidx.lifecycle.SavedStateHandle
-import androidx.test.core.app.ApplicationProvider
 import com.pocketshell.app.composer.PromptComposerViewModel.ApiKeyVault
 import com.pocketshell.app.composer.PromptComposerViewModel.RecordingState
 import com.pocketshell.app.diagnostics.installRecordingDiagnosticSink
@@ -792,16 +791,6 @@ class PromptComposerViewModelTest {
     }
 
     @Test
-    fun typedDraftIsEditableAndPreservedUntilCallerClearsIt() {
-        val vm = newVm()
-
-        vm.onDraftChange("first dictated idea")
-        vm.onDraftChange("first dictated idea with an edit")
-
-        assertEquals("first dictated idea with an edit", vm.uiState.value.draft)
-    }
-
-    @Test
     fun clearErrorRemovesBanner() {
         val vm = newVm()
         vm.surfacePermissionDenied()
@@ -1518,351 +1507,6 @@ class PromptComposerViewModelTest {
         assertEquals(PromptComposerViewModel.RecordingState.Idle, vm.uiState.value.recording)
     }
 
-    // -- Issue #746: discard draft + session scoping ------------------------
-
-    @Test
-    fun discardDraftClearsTextAttachmentsBannerAndSavedState() = runTest {
-        // Issue #746: the "Not sent. …or discard the draft." banner promised a
-        // Discard action that did not exist. discardDraft() must wipe the draft
-        // text, every staged attachment, the error/status banner, and the
-        // persisted SavedStateHandle keys so the next open is a clean slate.
-        val savedStateHandle = SavedStateHandle()
-        val vm = newVm(
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            savedStateHandle = savedStateHandle,
-        )
-        vm.onDraftChange("ничего не происходит")
-        vm.attachFiles(count = 1) {
-            Result.success(listOf("~/.pocketshell/attachments/host-1/shot.png"))
-        }
-        advanceUntilIdle()
-        // The draft + attachment chip are staged before the failed send.
-        assertTrue(vm.uiState.value.draft.isNotEmpty())
-        assertTrue(vm.uiState.value.attachments.isNotEmpty())
-        // Simulate the degraded-send "Not sent" banner. The "Not sent" state is a
-        // draft (+ since #872, the restored attachment tiles) + banner the user
-        // must be able to discard. Here we pass a bare SendRequest (clean draft
-        // defaults to text, no tiles) so the assertions below focus on discard.
-        vm.restoreFailedSend(
-            PromptComposerViewModel.SendRequest(text = "ничего не происходит", withEnter = true),
-        )
-        assertTrue(vm.uiState.value.draft.isNotEmpty())
-        assertNotNull(vm.uiState.value.error)
-
-        vm.discardDraft()
-
-        assertEquals("", vm.uiState.value.draft)
-        assertTrue(vm.uiState.value.attachments.isEmpty())
-        assertNull(vm.uiState.value.error)
-        // SavedStateHandle is wiped so a process-death recreate stays clean.
-        assertEquals("", savedStateHandle.get<String>(PromptComposerViewModel.KEY_DRAFT).orEmpty())
-        assertNull(savedStateHandle.get<String>(PromptComposerViewModel.KEY_DRAFT_OWNER))
-    }
-
-    @Test
-    fun discardDraftClearsStagedAttachmentChipsBeforeSend() = runTest {
-        // Issue #746: discard must also clear staged attachment chips that are
-        // still present (i.e. before a send folded them into the draft text).
-        val vm = newVm(samplerDispatcher = StandardTestDispatcher(testScheduler))
-        vm.onDraftChange("with a chip")
-        vm.attachFiles(count = 1) {
-            Result.success(listOf("~/.pocketshell/attachments/host-1/shot.png"))
-        }
-        advanceUntilIdle()
-        assertTrue(vm.uiState.value.attachments.isNotEmpty())
-
-        vm.discardDraft()
-
-        assertEquals("", vm.uiState.value.draft)
-        assertTrue(vm.uiState.value.attachments.isEmpty())
-    }
-
-    @Test
-    fun draftFromOneSessionDoesNotBleedIntoAnother() = runTest {
-        // Issue #746: a "Not sent" draft authored in session A must NOT appear
-        // when the composer is opened/focused in session B. Switching the
-        // target session discards the stale draft + banner.
-        val vm = newVm(samplerDispatcher = StandardTestDispatcher(testScheduler))
-        vm.onComposerTargetChanged("1/sessionA")
-        vm.restoreFailedSend(
-            PromptComposerViewModel.SendRequest(text = "draft for A", withEnter = true),
-        )
-        assertEquals("draft for A", vm.uiState.value.draft)
-        assertNotNull(vm.uiState.value.error)
-
-        // Switch to session B: the stale draft + banner are gone.
-        vm.onComposerTargetChanged("1/sessionB")
-        assertEquals("", vm.uiState.value.draft)
-        assertNull(vm.uiState.value.error)
-    }
-
-    @Test
-    fun draftSurvivesReturningToTheOwningSession() = runTest {
-        // Issue #746: scoping discards the draft only in OTHER sessions. While
-        // the composer stays on the session that authored the draft, the draft
-        // (and its retry banner) must be preserved so the user can resend.
-        val vm = newVm(samplerDispatcher = StandardTestDispatcher(testScheduler))
-        vm.onComposerTargetChanged("1/sessionA")
-        vm.restoreFailedSend(
-            PromptComposerViewModel.SendRequest(text = "draft for A", withEnter = true),
-        )
-        // Re-reporting the same target (e.g. a recomposition) is a no-op.
-        vm.onComposerTargetChanged("1/sessionA")
-        assertEquals("draft for A", vm.uiState.value.draft)
-        assertNotNull(vm.uiState.value.error)
-    }
-
-    @Test
-    fun processRestoredDraftIsAdoptedByFirstTargetNotDropped() = runTest {
-        // Issue #746: a draft restored from SavedStateHandle after process
-        // death has no owner stamp yet. The FIRST target the host reports must
-        // ADOPT it (the user is back where they typed it) rather than discard a
-        // legitimately-recovered prompt. A subsequent switch to a different
-        // session then discards it.
-        val savedStateHandle = SavedStateHandle()
-        savedStateHandle[PromptComposerViewModel.KEY_DRAFT] = "recovered draft"
-        val vm = newVm(
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            savedStateHandle = savedStateHandle,
-        )
-        assertEquals("recovered draft", vm.uiState.value.draft)
-
-        // First target after restore adopts the orphan draft.
-        vm.onComposerTargetChanged("1/sessionA")
-        assertEquals("recovered draft", vm.uiState.value.draft)
-
-        // A later switch to a different session discards it.
-        vm.onComposerTargetChanged("1/sessionB")
-        assertEquals("", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun emptyDraftSwitchJustReassignsTargetWithoutClearingFutureDrafts() = runTest {
-        // Issue #746: switching sessions with NO draft is a no-op clear, and a
-        // draft typed afterwards is owned by the now-current session so it
-        // persists on a same-session recomposition.
-        val vm = newVm(samplerDispatcher = StandardTestDispatcher(testScheduler))
-        vm.onComposerTargetChanged("1/sessionA")
-        vm.onComposerTargetChanged("1/sessionB")
-        vm.onDraftChange("typed in B")
-        // Same-session re-report keeps the draft.
-        vm.onComposerTargetChanged("1/sessionB")
-        assertEquals("typed in B", vm.uiState.value.draft)
-        // Switching away from B discards B's draft.
-        vm.onComposerTargetChanged("1/sessionA")
-        assertEquals("", vm.uiState.value.draft)
-    }
-
-    // -- Issue #832: durable per-session draft store ------------------------
-
-    @Test
-    fun failedSendDraftSurvivesSwitchAwayAndBack() = runTest {
-        // Issue #832 reproduction (the maintainer's exact dogfood scenario): an
-        // attachment/send fails mid-session (silent drop), the composer keeps
-        // the draft ("Your draft was kept"), the user switches to another
-        // session, then returns — and the draft MUST still be there. Before the
-        // durable per-session store this draft was discarded on the FIRST
-        // switch and was unrecoverable through the UI.
-        //
-        // RED without the fix: with the per-session store wired, the switch
-        // back used to find an empty draft (the old onComposerTargetChanged
-        // discarded it). GREEN: the store reloads session A's draft.
-        val store = InMemoryComposerDraftStore()
-        val vm = newVm(
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            composerDraftStore = store,
-        )
-        vm.onComposerTargetChanged("1/sessionA")
-
-        // The send into session A fails; the composer keeps the draft.
-        vm.restoreFailedSend(
-            PromptComposerViewModel.SendRequest(text = "draft for A", withEnter = true),
-        )
-        assertEquals("draft for A", vm.uiState.value.draft)
-        assertNotNull(vm.uiState.value.error)
-
-        // The user switches to session B: A's draft must NOT bleed into B.
-        vm.onComposerTargetChanged("1/sessionB")
-        assertEquals("", vm.uiState.value.draft)
-        assertNull(vm.uiState.value.error)
-
-        // The user returns to session A: the kept draft is restored.
-        vm.onComposerTargetChanged("1/sessionA")
-        assertEquals("draft for A", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun multipleSessionsEachKeepTheirOwnDurableDraft() = runTest {
-        // Issue #832 class-coverage: distinct drafts in three sessions all
-        // survive arbitrary switching; each session only ever shows its own.
-        val store = InMemoryComposerDraftStore()
-        val vm = newVm(
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            composerDraftStore = store,
-        )
-
-        vm.onComposerTargetChanged("1/a")
-        vm.onDraftChange("alpha draft")
-        vm.onComposerTargetChanged("1/b")
-        assertEquals("", vm.uiState.value.draft)
-        vm.onDraftChange("bravo draft")
-        vm.onComposerTargetChanged("2/c")
-        assertEquals("", vm.uiState.value.draft)
-        vm.onDraftChange("charlie draft")
-
-        // Cycle back through every session: each reloads its own draft.
-        vm.onComposerTargetChanged("1/a")
-        assertEquals("alpha draft", vm.uiState.value.draft)
-        vm.onComposerTargetChanged("2/c")
-        assertEquals("charlie draft", vm.uiState.value.draft)
-        vm.onComposerTargetChanged("1/b")
-        assertEquals("bravo draft", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun editingDraftAfterSwitchBackUpdatesTheRightSessionsStore() = runTest {
-        // Issue #832: an edit after returning to a session is persisted under
-        // THAT session, not the one we switched away from.
-        val store = InMemoryComposerDraftStore()
-        val vm = newVm(
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            composerDraftStore = store,
-        )
-        vm.onComposerTargetChanged("1/a")
-        vm.onDraftChange("a v1")
-        vm.onComposerTargetChanged("1/b")
-        vm.onComposerTargetChanged("1/a")
-        vm.onDraftChange("a v2")
-
-        assertEquals("a v2", store.load("1/a"))
-        assertNull(store.load("1/b"))
-    }
-
-    @Test
-    fun sharedPrefsDraftWritesAreScheduledOffCallerThreadButVisibleToSessionSwitches() = runTest {
-        val dispatcher = StandardTestDispatcher(testScheduler)
-        val store = SharedPrefsComposerDraftStore(ApplicationProvider.getApplicationContext())
-        val target = "test/off-main-${System.nanoTime()}"
-        val vm = newVm(
-            samplerDispatcher = dispatcher,
-            composerDraftStore = store,
-        )
-        vm.onComposerTargetChanged(target)
-
-        vm.onDraftChange("first")
-        vm.onDraftChange("second")
-
-        assertEquals("second", vm.uiState.value.draft)
-        assertNull(store.load(target))
-
-        vm.onComposerTargetChanged("$target/other")
-        vm.onComposerTargetChanged(target)
-
-        assertEquals("second", vm.uiState.value.draft)
-        assertNull(store.load(target))
-
-        advanceUntilIdle()
-
-        assertEquals("second", store.load(target))
-    }
-
-    @Test
-    fun clearDraftWinsOverPendingSharedPrefsDraftSave() = runTest {
-        val dispatcher = StandardTestDispatcher(testScheduler)
-        val store = SharedPrefsComposerDraftStore(ApplicationProvider.getApplicationContext())
-        val target = "test/clear-wins-${System.nanoTime()}"
-        val vm = newVm(
-            samplerDispatcher = dispatcher,
-            composerDraftStore = store,
-        )
-        vm.onComposerTargetChanged(target)
-
-        vm.onDraftChange("stale draft")
-        vm.discardDraft()
-
-        assertEquals("", vm.uiState.value.draft)
-        advanceUntilIdle()
-
-        assertNull(store.load(target))
-    }
-
-    @Test
-    fun discardClearsTheDurableSlotSoSwitchBackIsEmpty() = runTest {
-        // Issue #832: Discard is the explicit "throw it away" control — it must
-        // drop the durable slot too, otherwise a switch back resurrects the
-        // draft the user just discarded.
-        val store = InMemoryComposerDraftStore()
-        val vm = newVm(
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            composerDraftStore = store,
-        )
-        vm.onComposerTargetChanged("1/a")
-        vm.restoreFailedSend(
-            PromptComposerViewModel.SendRequest(text = "stale prompt", withEnter = true),
-        )
-        assertEquals("stale prompt", store.load("1/a"))
-
-        vm.discardDraft()
-        assertNull(store.load("1/a"))
-
-        // Switch away and back: the discarded draft does not come back.
-        vm.onComposerTargetChanged("1/b")
-        vm.onComposerTargetChanged("1/a")
-        assertEquals("", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun deliveredSendClearsTheDurableSlot() = runTest {
-        // Issue #832: a confirmed delivery wipes the draft AND its durable
-        // slot, so the next time the user lands on that session the composer
-        // is a clean slate (not a resurrected just-sent prompt).
-        val store = InMemoryComposerDraftStore()
-        val vm = newVm(
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            composerDraftStore = store,
-        )
-        vm.onComposerTargetChanged("1/a")
-        vm.onDraftChange("about to send")
-        assertEquals("about to send", store.load("1/a"))
-
-        vm.markSendDelivered()
-        assertNull(store.load("1/a"))
-
-        vm.onComposerTargetChanged("1/b")
-        vm.onComposerTargetChanged("1/a")
-        assertEquals("", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun restoreFailedSendUsesRequestTargetWhenCurrentComposerMoved() = runTest {
-        // Issue #900: a send can resolve after the user has switched sessions.
-        // The failed draft belongs to the tap-time session carried by the request,
-        // not whatever composer target is current when the callback returns.
-        val store = InMemoryComposerDraftStore()
-        val vm = newVm(
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            composerDraftStore = store,
-        )
-        vm.onComposerTargetChanged("1/a")
-        val request = PromptComposerViewModel.SendRequest(
-            text = "draft for A",
-            withEnter = true,
-            sendTarget = PromptComposerViewModel.SendTargetSnapshot(sessionKey = "1/a"),
-        )
-
-        vm.onComposerTargetChanged("1/b")
-        vm.onDraftChange("draft for B")
-        vm.restoreFailedSend(request)
-
-        assertEquals("draft for B", vm.uiState.value.draft)
-        assertNull(vm.uiState.value.error)
-        assertEquals("draft for A", store.load("1/a"))
-        assertEquals("draft for B", store.load("1/b"))
-
-        vm.onComposerTargetChanged("1/a")
-        assertEquals("draft for A", vm.uiState.value.draft)
-    }
-
     @Test
     fun composerTargetChangedRefreshesOutboundQueueForCurrentSession() = runTest {
         val queue = InMemoryOutboundQueueStore()
@@ -1949,36 +1593,6 @@ class PromptComposerViewModelTest {
             listOf(inFlight.id, uploading.id),
             vm.outboundQueueItems.value.map { it.id },
         )
-    }
-
-    @Test
-    fun deliveredSendUsesRequestTargetWhenCurrentComposerMoved() = runTest {
-        // Issue #900: the success path has the same stale-callback risk as
-        // restore. A delayed success for A must clear A's durable draft without
-        // clearing B's currently visible draft.
-        val store = InMemoryComposerDraftStore()
-        val vm = newVm(
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-            composerDraftStore = store,
-        )
-        vm.onComposerTargetChanged("1/a")
-        vm.onDraftChange("about to send")
-        val request = PromptComposerViewModel.SendRequest(
-            text = "about to send",
-            withEnter = true,
-            sendTarget = PromptComposerViewModel.SendTargetSnapshot(sessionKey = "1/a"),
-        )
-
-        vm.onComposerTargetChanged("1/b")
-        vm.onDraftChange("draft for B")
-        vm.markSendDelivered(request)
-
-        assertNull(store.load("1/a"))
-        assertEquals("draft for B", vm.uiState.value.draft)
-        assertEquals("draft for B", store.load("1/b"))
-
-        vm.onComposerTargetChanged("1/a")
-        assertEquals("", vm.uiState.value.draft)
     }
 
     @Test
