+++ scripts/component-drift-baseline.txt
@@ -21,7 +21,7 @@ app/src/main/java/com/pocketshell/app/snippets/SnippetPickerSheet.kt 1
 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionAuxiliaryModals.kt 6
 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionConnectionChrome.kt 4
 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionOverlays.kt 5
-app/src/main/java/com/pocketshell/app/tmux/TmuxSessionScreen.kt 4
+app/src/main/java/com/pocketshell/app/tmux/TmuxSessionScreen.kt 1
 shared/ui-kit/src/main/java/com/pocketshell/uikit/components/ConfirmDialog.kt 1
 shared/ui-kit/src/main/java/com/pocketshell/uikit/components/FormDialog.kt 1
 shared/ui-kit/src/main/java/com/pocketshell/uikit/components/HostCard.kt 2
